requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
opencv-python==4.5.3.56
albumentations==1.4.0
efficientnet-pytorch==0.7.1
ultralytics==8.2.63
scikit-learn==1.2.2
pandas==2.1.4
numpy==1.26.4
matplotlib==3.9.1
seaborn==0.13.2
Pillow==10.4.0
tqdm==4.65.0
requests==2.32.3
s3fs==2023.10.0
huggingface-hub==0.17.3
